+++ src/App.tsx
@@ -1,8 +1,7 @@
 import { useState } from 'react';
 import { SnowCanvas } from './components/ui/SnowCanvas';
 import { Navbar } from './components/layout/Navbar';
-import { HeroSection } from './components/sections/HeroSection';
-import { DataSection } from './components/sections/DataSection';
+import { TopSection } from './components/sections/TopSection';
 import { GallerySection } from './components/sections/GallerySection';
 import { GuestbookSection } from './components/sections/GuestbookSection';
 import { Footer } from './components/sections/Footer';
@@ -39,8 +38,7 @@ function App() {
 
       {/* Layer 2: Page content */}
       <main>
-        <HeroSection onCallYoshinon={openWelcome} />
-        <DataSection />
+        <TopSection />
         <GallerySection />
         <GuestbookSection />
       </main>